📝 docs(README.md): update introduction section with additional information and links
The introduction section of the README.md file has been updated to provide more information about the Tailwind Palette Python package. It now includes a link to the original [python-tailwind-colors](https://github.com/dostuffthatmatters/python-tailwind-colors) library and a link to the Tailwind CSS color palette documentation on the official Tailwind CSS website. This update aims to improve the user experience and provide additional resources for users.